better subdir checking
